Project/requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
async-timeout==5.0.1
attrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.1.8
clip @ git+https://github.com/openai/CLIP.git@d05afc436d78f1c48dc0dbf8e5980a9d471f35f6
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
filelock==3.19.1
flatbuffers==25.12.19
fonttools==4.60.2
frozenlist==1.8.0
fsspec==2025.10.0
ftfy==6.3.1
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.7
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.9.4
mediapipe==0.10.33
mpmath==1.3.0
multidict==6.7.1
networkx==3.2.1
nltk==3.9.2
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu11==8.7.0.84
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu11==2.19.3
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.8.90
opencv-python==4.8.1.78
packaging==26.0
pandas==2.3.3
pillow==11.3.0
portalocker==3.2.0
propcache==0.4.1
psutil==7.2.2
pycparser==2.23
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
regex==2026.1.15
requests==2.32.5
rouge_score==0.1.2
sacrebleu==2.6.0
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sounddevice==0.5.5
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
# torch==2.2.0+cu118
# torch-geometric==2.6.1
# torch_cluster==1.6.3+pt22cu118
# torch_scatter==2.1.2+pt22cu118
# torch_sparse==0.6.18+pt22cu118
# torch_spline_conv==1.2.2+pt22cu118
# torchaudio==2.2.0+cu118
# torchvision==0.17.0+cu118
tqdm==4.67.3
triton==2.2.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wcwidth==0.6.0
yarl==1.22.0
zipp==3.23.0
